backend/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
psycopg2==2.9.3
django-money==3.0.0
gunicorn==20.1.0
uvicorn[standard]==0.18.3
djangorestframework-simplejwt[crypto]==5.2.1
Pillow==9.2.0
django-cors-headers==3.13.0
django-countries==7.3.2